+++ mrz-3-edition/chapters/5/algorithms/2/jacobiEigenCycle.m
@@ -1,22 +1,19 @@
-function [x, numberOfIterations, derivativeZero] = newton(f, fDerivative, startPoint, tolerance, maxIterations)
-% NEWTON: Finds a solution of f(x) = 0 with Newton method.
+function [A, V, k, nr] = jacobiEigenCycle(A, toll, nmax)
+% JACOBIEIGEN: finds eigenvalues and eigenvectors of matrix using cyclic
+% Jacobi method
 %
-%  [x, numberOfIterations, derivativeZero] = newton(f, fDerivative, startPoint, tolerance, maxIterations)
+%  [A, V, k, nr] = jacobiEigen(A, toll, nmax)
 %
 % Input:
-% f - 'f' function in the equation 'f(x) = 0'
-% fDerivative - derivative of f
-% startPoint - starting point of method
-% tolerance - epsilon at which stop method (i.e when |f(xn) - 0| <
-%             epsilon)
-% maxIterations - max number of iterations to execute
+% A - simmetric matrix
+% toll - min tolerance of output error
+% nmax - max number of iterations (rotations)
 %
 % Output:
-% x - approximation of solution of 'f(x) = 0'
-% numberOfIterations - number of iterations executed before getting
-%                      solution
-% derivativeZero - boolean to check whether function has stopped or not
-%                  because of a f'(x) = 0
+% A - input matrix with eigenvalues on diagonal
+% V - matrix: each column is an eigenvector of A
+% k - number of full cycles done
+% nr - number of rotations done
 
 % Copyright 2017 Stefano Fogarollo
 %
@@ -32,18 +29,72 @@ function [x, numberOfIterations, derivativeZero] = newton(f, fDerivative, startP
 % See the License for the specific language governing permissions and
 % limitations under the License.
 
-numberOfIterations = 0;
-derivativeZero = false;
-x = startPoint;
-deltaDiff = tolerance * 2;  % initialize diff
-while deltaDiff >= tolerance && numberOfIterations < maxIterations && ~derivativeZero
-    derivativeValue = feval(fDerivative, x);
-    if derivativeValue == 0
-        derivativeZero = true;
-    else
-        deltaDiff = - feval(f, x) / derivativeValue;
-        x = x + deltaDiff;
-        deltaDiff = abs(deltaDiff);
+n = size(A, 1);  % number of rows
+V = eye(n);
+k = 0;
+nr = 0;
+tolla = sqrt(toll / (n ^ 2 - n));
+
+%% Calculate initial tolerance
+testToll = 0;
+for q = 1 : n
+    for p = 1 : q - 1
+        testToll = testToll + abs(A(p, q)) ^ 2;
     end
-    numberOfIterations  = numberOfIterations + 1;  % increase counter
+end
+testToll = 2 * testToll;
+
+while testToll > toll && nr < nmax
+    k = k + 1;
+    for p = 1 : n - 1
+        for q = p + 1 : n
+            if abs(A(p, q)) >= tolla && testToll >= toll
+                %% Perform rotation
+                u = (A(p, p) - A(q, q)) / (2 * A(p, q));
+                t = 1 / (abs(u) + sqrt(1 + u ^ 2));
+                if u < 0
+                    t = -t;
+                end
+                c = 1 / sqrt(1 + t ^ 2);
+                s = t * c;
+                tau = s / (1 + c);
+                A(p, p) = A(p, p) + t * A(p, q);
+                A(q, q) = A(q, q) - t * A(p, q);
+                testToll = testToll - 2 * A(p, q) ^ 2;
+                A(p, q) = 0;
+
+                for i = 1 : p - 1
+                    tmp = A(i, p);
+                    A(i, p) = A(i, p) + s * (A(i, q) - tau * A(i, p));
+                    A(i, q) = A(i, q) - s * (tmp + tau * A(i, q));
+                end
+
+                for i = p + 1 : q - 1
+                    tmp = A(p, i);
+                    A(p, i) = A(p, i) + s * (A(i, q) - tau * A(p, i));
+                    A(i, q) = A(i, q) - s * (tmp + tau * A(i, q));
+                end
+
+                for i = q + 1 : n
+                    tmp = A(p, i);
+                    A(p, i) = A(p, i) + s * (A(q, i) - tau * A(p, i));
+                    A(q, i) = A(q, i) - s * (tmp + tau * A(q, i));
+                end
+
+                for i = 1 : n
+                    tmp = V(i, p);
+                    V(i, p) = V(i, p) + s * (V(i, q) - tau * V(i, p));
+                    V(i, q) = V(i, q) - s * (tmp + tau * V(i, q));
+                end
+
+                for j = 1 : n - 1
+                    for i = j + 1 : n
+                        A(i, j) = A(j, i);
+                    end
+                end
+
+                nr = nr + 1;  % update rotation counter
+            end
+        end
+    end
 end